and
I guess it’s okay to use the subsequent lines; the main point is to write a story in the first line so it comes out forward though I write it backwards.  By the way, I only have the interview written, and a possible last line of the story, but NOT a last paragraph nor arc nor anything substantial of the story itself.

I used gitx, not quotes to write this long commit message.